improve performance of blurbackground loading
